[dagster-shared] Move core telemetry code to dagster-shared
